API

NestedGraphMakie.ngraphplot! — Method

ngraphplot(ng::NestedGraph)
ngraphplot!(ax, ng::NestedGraph)

Plots a NestedGraph. Actually decorates GraphMakie.graphplot with some extra features.

Attributes

colors=automatic: Vector of colors for the 1st-level subgraphs

multilayer=false: Plot 1st-level subgraphs in a multilayer fashion

multilayer_dist=automatic: how far away should the layer of the graph be

fold_graphs=[Int[]]: choose which graphs not to unfold (experimental)

show_subgraph_regions=false depicts all subgraphs in the plot

NestedGraphMakie.ngraphplot — Method

ngraphplot(ng::NestedGraph)
ngraphplot!(ax, ng::NestedGraph)

Plots a NestedGraph. Actually decorates GraphMakie.graphplot with some extra features.

Attributes

colors=automatic: Vector of colors for the 1st-level subgraphs

multilayer=false: Plot 1st-level subgraphs in a multilayer fashion

multilayer_dist=automatic: how far away should the layer of the graph be

fold_graphs=[Int[]]: choose which graphs not to unfold (experimental)

show_subgraph_regions=false depicts all subgraphs in the plot

Index

NestedGraphMakie.ngraphplot

NestedGraphMakie.ngraphplot!